rpy/requirements.txt
picamera==1.13
flask==1.0.2
Pillow==5.4.1
numpy==1.16.2
